requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
git-filter-repo==2.47.0
idna==3.11
iniconfig==2.3.0
loguru==0.7.3
lxml==6.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
